--- app/src/main/java/com/example/movietheater/base/Event.kt
@@ -1,3 +1,0 @@
-package com.example.movietheater.base
-
-interface Event
