+++ app/src/main/java/com/example/movietheater/base/viewmodel/Event.kt
@@ -1,0 +1,3 @@
+package com.example.movietheater.base.viewmodel
+
+interface Event
