+++ app/src/main/java/com/example/movietheater/base/viewmodel/Status.kt
@@ -1,4 +1,4 @@
-package com.example.movietheater.base
+package com.example.movietheater.base.viewmodel
 
 enum class Status {
     CONTENT,
